fix(test): v0.10.0 pg_regress — ingest ::smallint + regenerate 4 .out fixtures
CI installcheck failed 4/30, blocking the entire publish pipeline. Root cause:
(1) provenance_gate T6/T7 called pgmnemo.ingest() with bare integer importance — PG
does not implicitly cast integer->smallint in function resolution -> 'function does
not exist'. Fixed with 3::smallint (canonical call).
(2) recall_fast / navigate_dispatch / selective_embedding / provenance_gate .out
fixtures were not generated by pg_regress (trailing whitespace stripped, missing
command-echo lines). Regenerated all 4 from a clean pgvector/pg17 'make installcheck'.
Result: All 30 tests pass on a clean contour.

+++ extension/sql/provenance_gate.sql
@@ -109,7 +109,7 @@ SELECT EXISTS (
 SELECT pgmnemo.ingest(
     'tc_pggate', NULL, 'gate_ingest_sha',
     'T6: ingest with commit_sha stamps verified_at.',
-    3, NULL, 'sha-t6-abc', NULL, '{}'::jsonb
+    3::smallint, NULL, 'sha-t6-abc', NULL, '{}'::jsonb
 ) IS NOT NULL AS t6_ingest_id;
 
 SELECT verified_at IS NOT NULL AS t6_verified
@@ -127,7 +127,7 @@ SET pgmnemo.gate_strict = 'off';
 SELECT pgmnemo.ingest(
     'tc_pggate', NULL, 'gate_ingest_ghost',
     'T7: ingest without provenance yields ghost lesson.',
-    3, NULL, NULL, NULL, '{}'::jsonb
+    3::smallint, NULL, NULL, NULL, '{}'::jsonb
 ) IS NOT NULL AS t7_ingest_id;
 
 SELECT verified_at IS NULL AS t7_ghost
